tests/submission/W3C/domhtml/HTMLObjectElement11.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-object-element.html#dom-object-type" rel="help">
<link rel="author" title="Netscape">
</head>
<body>

<object data="/support/html/sample.html" height="80" type="text/html" width="550"></object>


<script>
var domtest = async_test("The IDL attribute type must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtype = testNode.type;

      assert_equals(vtype,"text/html");
       
    domtest.done();
}));

</script>
<p>Content type for data downloaded via data attribute.
The value of attribute type of the object element is read and checked against the expected value.</p>
<div id="log"></div>
</body></html>
